+++ packages/experience-builder-sdk/src/blocks/preview/CompositionBlock.tsx
@@ -78,27 +78,23 @@ export const CompositionBlock = ({
   }, [isAssembly, node.definitionId]);
 
   const nodeProps = useMemo(() => {
+    // In SSR, we store the className under breakpoints[0] which is resolved here to the actual string
+    const cfSsrClassNameValues = node.variables.cfSsrClassName as DesignValue | undefined;
+    const cfSsrClassName = resolveDesignValue(
+      cfSsrClassNameValues?.valuesByBreakpoint,
+      'cfSsrClassName',
+    );
+
     // Don't enrich the assembly wrapper node with props
     if (!componentRegistration || isAssembly) {
-      return {
-        cfSsrClassName: node.variables.cfSsrClassName
-          ? resolveDesignValue(
-              (node.variables.cfSsrClassName as DesignValue).valuesByBreakpoint,
-              'cfSsrClassName',
-            )
-          : undefined,
-      };
+      return { cfSsrClassName };
     }
 
     const propMap: Record<string, PrimitiveValue> = {
-      cfSsrClassName: node.id
-        ? getPatternChildNodeClassName?.(node.id)
-        : node.variables.cfSsrClassName
-          ? resolveDesignValue(
-              (node.variables.cfSsrClassName as DesignValue).valuesByBreakpoint,
-              'cfSsrClassName',
-            )
-          : undefined,
+      cfSsrClassName:
+        node.id && getPatternChildNodeClassName
+          ? getPatternChildNodeClassName(node.id)
+          : cfSsrClassName,
     };
 
     const props = Object.entries(componentRegistration.definition.variables).reduce(
@@ -182,11 +178,13 @@ export const CompositionBlock = ({
 
     return props;
   }, [
+    resolveDesignValue,
+    node.variables,
+    node.id,
+    node.children,
     componentRegistration,
     isAssembly,
-    node.children,
-    node.variables,
-    resolveDesignValue,
+    getPatternChildNodeClassName,
     entityStore,
     hyperlinkPattern,
     locale,
@@ -200,17 +198,13 @@ export const CompositionBlock = ({
 
   const { component } = componentRegistration;
 
+  // Retrieves the CSS class name for a given child node ID.
   const _getPatternChildNodeClassName = (childNodeId: string) => {
     if (isAssembly) {
       // @ts-expect-error -- property cfSsrClassName is a map (id to classNames) that is added during rendering in ssrStyles
-      const classesForNode = node.variables.cfSsrClassName[childNodeId];
-      if (classesForNode) {
-        return resolveDesignValue(
-          (classesForNode as DesignValue).valuesByBreakpoint,
-          'cfSsrClassName',
-        ) as string;
-      }
-      return;
+      const classesForNode: DesignValue | undefined = node.variables.cfSsrClassName?.[childNodeId];
+      if (!classesForNode) return undefined;
+      return resolveDesignValue(classesForNode.valuesByBreakpoint, 'cfSsrClassName') as string;
     }
     return getPatternChildNodeClassName?.(childNodeId);
   };
@@ -220,7 +214,11 @@ export const CompositionBlock = ({
       ? node.children.map((childNode: ComponentTreeNode, index) => {
           return (
             <CompositionBlock
-              getPatternChildNodeClassName={_getPatternChildNodeClassName}
+              getPatternChildNodeClassName={
+                isAssembly || getPatternChildNodeClassName
+                  ? _getPatternChildNodeClassName
+                  : undefined
+              }
               node={childNode}
               key={index}
               locale={locale}