+++ packages/experience-builder-sdk/src/hooks/useBreakpoints.ts
@@ -1,9 +1,4 @@
-import type {
-  ValuesByBreakpoint,
-  Breakpoint,
-  PrimitiveValue,
-  ResolveDesignValueType,
-} from '@contentful/experiences-core/types';
+import type { Breakpoint, ResolveDesignValueType } from '@contentful/experiences-core/types';
 import {
   mediaQueryMatcher,
   getFallbackBreakpointIndex,
@@ -55,7 +50,7 @@ export const useBreakpoints = (breakpoints: Breakpoint[]) => {
   );
 
   const resolveDesignValue: ResolveDesignValueType = useCallback(
-    (valuesByBreakpoint: ValuesByBreakpoint, variableName: string): PrimitiveValue => {
+    (valuesByBreakpoint, variableName) => {
       return getValueForBreakpoint(
         valuesByBreakpoint,
         breakpoints,