+++ packages/experience-builder-sdk/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.27.1](https://github.com/contentful/experience-builder/compare/v1.27.0...v1.27.1) (2024-12-19)
+
+**Note:** Version bump only for package @contentful/experiences-sdk-react
+
+## [1.27.0](https://github.com/contentful/experience-builder/compare/v1.27.0-beta.0...v1.27.0) (2024-12-17)
+
+**Note:** Version bump only for package @contentful/experiences-sdk-react
+
 ## [1.26.0](https://github.com/contentful/experience-builder/compare/v1.26.0-beta.0...v1.26.0) (2024-12-10)
 
 **Note:** Version bump only for package @contentful/experiences-sdk-react